refactor(gui): extract components
